Validate distribution->docker parameter, restrict dynamic context in Spell Expressions
