+++ component-resolver-core/src/main/java/org/octopusden/octopus/escrow/dto/EscrowExpressionContext.java
@@ -6,24 +6,19 @@ import org.octopusden.releng.versions.NumericVersionFactory;
 import java.util.Map;
 import java.util.function.Function;
 
-public class EscrowExpressionContext {
+public class EscrowExpressionContext extends SimpleExpressionContext {
     private final Map<String, String> env = System.getenv();
     private final String fileName;
-    private final String version;
-    private String component;
-    private String baseDir;
-    private IVersionInfo versionInfo;
+    private final String baseDir;
 
     public EscrowExpressionContext(String component, String version, String fileName, Function<String, IVersionInfo> versionItemExtractor) {
-        this.component = component;
+        super(component, version, versionItemExtractor);
         this.fileName = fileName;
-        this.version = version;
         baseDir = System.getProperty("user.dir");
-        versionInfo = versionItemExtractor.apply(version);
     }
 
     public EscrowExpressionContext(String component, String version, String fileName, NumericVersionFactory numericVersionFactory) {
-        this(component, version, fileName, componentVersion -> numericVersionFactory.create(componentVersion));
+        this(component, version, fileName, numericVersionFactory::create);
     }
 
     //TODO: remove? is it used?
@@ -31,41 +26,11 @@ public class EscrowExpressionContext {
         return fileName;
     }
 
-    public String getVersion() {
-        return version;
-    }
-
     public Map<String, String> getEnv() {
         return env;
-    }
-
-    public String getComponent() {
-        return component;
     }
 
     public String getBaseDir() {
         return baseDir;
     }
-
-    public int getMajor() {
-        return versionInfo.getMajor();
-    }
-
-    public int getMinor() {
-        return versionInfo.getMinor();
-    }
-
-    public int getService() {
-        return versionInfo.getService();
-    }
-
-    public int getFix() {
-        return versionInfo.getFix();
-    }
-
-    public int getBuild() {
-        return versionInfo.getBuildNumber();
-    }
-
-    //TODO: support major02, minor02, service02, fix02, fix04, build02, build04 etc. (like in org.octopusden.releng.versions.KotlinVersionFormatter.PREDEFINED_COMPONENT_VARIABLES_LIST)
 }
